Add unique index on login which is case insensitive
A cooler solution would have been to use the citext extension
and data type for the login field, however the extension require
additional packages to be install on the database host.
Since we aren't controlling the host environment where the Postgres
database runs, I opted for the index creation instead even
though it will require a slightly fiddlier implementation in the
users service.