Protect against join timing attacks
We mark subqueries vulnerable to join timing attacks during compilation and
we then make sure, during offloading, that they return at least one row
that doesn't match anything else in the data set.
We can't use `NULL` values for the invalid row, as some optimizers can detect
it will never match the join filters and drop it prematurely, so we choose the
set of values with the highest chance of not matching anything else in the data set.