Fix usages of global.trans
It seems we've been misusing `:global.trans`. The second element of the
`id` tuple is the id of the requester of the lock. If the id is the same
as that of a process that already possesses the lock, then the lock is
granted again. This is not what we want in the places where we have been
using it. The problem and solution is demonstrated by these programs:

```elixir
defmodule A do
  def in_trans() do
    :global.trans({__MODULE__, :a}, fn ->
      IO.puts("HELLO")
      :timer.sleep(1000)
      IO.puts("WORLD")
    end)
  end
end

task = Task.async(fn -> A.in_trans() end)
A.in_trans()
Task.await(task)
```

This one prints:

```
HELLO
HELLO
WORLD
WORLD
```

While this one:

```elixir
defmodule A do
  def in_trans() do
    :global.trans({__MODULE__, self()}, fn ->
      IO.puts("HELLO")
      :timer.sleep(1000)
      IO.puts("WORLD")
    end)
  end
end

task = Task.async(fn -> A.in_trans() end)
A.in_trans()
Task.await(task)
```

Prints:

```
HELLO
WORLD
HELLO
WORLD
```