Workaround restriction crashing on oracle
It seems for oracle queries in particular the restriction code tried to
resolve a subquery column that wasn't in the title list of the subquery.
This might or might not be a problem in general, however this piece of
code only requires the information if a given expression is a raw
database column or not.