Disable test triggering bug in Drill
The grouping tests seem to be failing because of a bug in drill, or
perhaps due to some other problem, but one that doesn't get properly
reported by drill due to a bug. Instead of producing an understandable
error message drill has this in the logs:

```
[Error Id: 4fbf13e3-2f50-4517-af44-071b1ceaf513 on 485cb1d0d163:31010]
org.apache.drill.common.exceptions.UserException: SYSTEM ERROR: CyclicMetadataException

Please, refer to logs for more information.

[Error Id: 4fbf13e3-2f50-4517-af44-071b1ceaf513 on 485cb1d0d163:31010]
        at org.apache.drill.common.exceptions.UserException$Builder.build(UserException.java:633) ~[drill-common-1.15.0.jar:1.15.0]
        at org.apache.drill.exec.work.foreman.Foreman$ForemanResult.close(Foreman.java:779) [drill-java-exec-1.15.0.jar:1.15.0]
        at org.apache.drill.exec.work.foreman.QueryStateProcessor.checkCommonStates(QueryStateProcessor.java:325) [drill-java-exec-1.15.0.jar:1.15.0]
        ...
```