Adding more test scaffolds
There are many paths we can go through our system to
manage user<--->group associations, as well as
group<--->data source associations.

A user can be added to a group, or a group to a user.
Likewise a data source can be added to a group, or a group to a data source.