Update test for new postgrex
It seems the new `postgrex` does not immediately discover that
credentials are invalid. I compared with `postgrex` 0.13 connecting to a
real postgres server and in that case it also doesn't immediately return
`{:error, _}` on connecting, so I'm not sure which behaviour is correct.
In any case, what we want to test is if it's possible to query, so this
test should be OK.