Compute consistent noise for grouping sets with stats-based anonymization
When using statistics-based anonymization, we offload computation of
grouping sets to the database. This requires more code and query emulation
for data sources not supporting the feature, but we will get better and
consistent results in exchange, since we won't have to do partial aggregation
of groups inside the cloak anymore.