Keep grouping columns as non-synthetic in the anonymized query
When using statistics-based anonymization, grouping columns were incorrectly
marked as synthetic in the top anonymization query and they were left as they
were in the synthetic UID-grouping subquery. This resulted in incorrect generation
of noise layers for grouping sets.

By inverting this state, we get correct creation of noise layers in all cases.

Unfortunately, this has the side-effect of breaking the reporting of query features,
as grouping is offloaded to the inner queries and the columns are marked as synthetic.
This will be fixed once we implement #3265.