Exit MongoCleanup if already run
This is a bit hacky, but they have already run the tool multiple times
over the same data twice, and debugging the problem is very
time-consuming, because it requires screensharing with them. A fuller
solution would somehow store the state of every object
(processed/unprocessed), allowing stopping/restarting the tool,
but that's more involved.