Add retries and wait for retries
We have some flaky acceptance tests that fail from time
to time. It is uneccessary and boring to have to restart
these tests, when they should work under normal circumstances.

This commit sets a number of retries to be attempted, in
the hopes that it will eliminate some of these issues.