Drop characters that are not valid UTF-8
TeamBank has seen the problem in practise where characters that aren't
invalid UTF-8, cause the MongoDB driver we are using to barf when
writing back the un-encrypted data. We circumvent this problem by
dropping everything that isn't a valid UTF-8 character. This will cause
minor dataloss, but at least leaves the remaining data available for
analysis.

I validated the solution by appending the offending character (0xDF13)
to the end of the decrypted string. This produced a stack trace for all
intents and purposes identical to the one sent to us by TeamBank.