Consistently use length and substring in Oracle
We were using the ____C variants of string functions in some places
in the Oracle driver. These give us unicode character lengths and
substrings respectively. Or rather give us the length and substrings
on the assumption that the underlying data actually is unicode,
whether or not it is.

The "undecorated" versions of the functions (length and substr)
give you the right values depending on the character set the database
is encoded to store. I.e. if the table is stored to hold unicode values,
then the unicode lengths and types will be returned, if however the
values use the western characterset ascii encoded then the right
character counts and substrings will also be returned.

One caveat here is what happens if a column is for example nvarchar
and thereby does hold unicode data whereas the database as a whole
does not? It's unclear what version of the function will be used in these
cases...