Use discovery query which does not execute views
If a complex view is being used rather than a regular table,
Oracle does not understand that the WHERE-clause of 0=1
will always fail. As a result it will compute the entire view
before not returning any data. Depending on the complexity
of the view this might take a significant amount of time to
complete.

This alternative query informs Oracle that no data is needed.
rownum < 0 or OFFSET 0 are not individually sufficient.
Telefonica derived at this as a good query through trial and error.