Exclude `null` input values from `count` statistics.
The `count` aggregator will always return a non-null result, even when
the entire input is null. This will distort the computed per-user
statistics, since users that do not contribute to the output at all
will still be included in them. We avoid this problem by switching to
a conditional `sum` when counting nullable values.