Improve join-timing attack protection.
The previous algorithm only worked for very simple cases. It didn't
detect filters in subqueries and it didn't properly protect against
join-timing attacks with more than 2 branches.

Unfortunately, we can't keep generating a dummy row in case a query
returns nothing, as that row won't match anything in subsequent joins,
making the attack successful. Instead, we now randomly keep a row when
the protected query returns nothing, and rely on the low-count filter
to ignore the result.