Remove semaphore in LifeCycle
We used a jobs queue in the life cycle state changing machinery to
prevent too many concurrent changes starving the rest of the application
for available Postgres connections. In hindsight this seems like
uneccessary over-engineering. We already have a pool of connections
available and these database updates are all quite quick and short
as it is.

Furthermore the jobs queue was not without problems itself. Job queue
failures elsewhere in the system caused a deadlock in this part of the
system, making everything brittle and unreliable.