Fix: link from query back to activity monitor
In the activity monitor in the admin section you can view details about a specific
query. In that query view there is a breadcrumb navigation menu which leads
offers a link back to the activity monitor. This link was incorrectly
linking to the main admin page, rather than the activity monitor.