Upgrades to phoenix 1.5 and adds Live-dashboard (#4484)
This also upgrades various related dependencies. Out of these the most notable is upgrading ranch from 1.3 to 1.7, which breaks the undocumented API we were previously using. This commit also changes that to use the documented way to manage an embedded ranch server.

This commit also paves the way for live_view related functionality.