Move native users out of LDAP groups
This is a one-off migration in case a system is in a broken state where a
native user is in an LDAP group. The other changes made in this pull-
request make it impossible to manually move a native user out of an LDAP
group. If we don't move them out for the sys admin, then there is no actual
way for the sys admin to do it for themself!