List query states in deterministic order
When there are too many queries for us to sensibly show in the activity
monitor, we revert to showing a subset of the queries along with a table
that shows how many queries there are in a given state. Up until now the
states table would constantly change in a way that made it impossible to
really digest. New items would be added, other would vanish, and the
order was inconsistent and always changing.

This commit uses the knowledge we have of the states that exist in the system
and always renders them all (whether or not there are queries in the
particular state). This ensures that the table remains fixed, and you
get to see how the queries flow through the system.