Simplify and fix tests
There was a lot of unneccessary and redundant testing.
For example we were extensively testing the previous version,
where all we really about is whether things work well after
an upgrade. Presumably the large scale tests were already
run prior to the previous release being made.

As we are blocking queries containing a `*`, the Tableu
tests no longer worked. The previous version contained a
section forspecifying a view in Tableau. This is no longer possible.