Make 1 month window default in audit log
At present it's 24 hours, which for systems tha are slightly less busy
is entirely useless. It was an OK choice in the past where the filtering
happened on the client side, was limited to a 1000 records in any case,
and you had to make good use of time slizing to get at the data you wanted,
but now that we have pagination, do processing server side, it seems rather moot.

Closes #4710