Add index on queries [user_id, context]
When, as a logged in user, you hit the root path of our app
we look up the last query you ran through the web interface
and redirect you to that particular data source page for querying.

The query issued filters on the user id and the context of the query.
While we had an index on the user id, we do not have an index on
the context. The resulting query ended up doing a linear search
through all the queries of the user:

```
Limit  (cost=2674.66..2674.66 rows=1 width=949)
  ->  Sort  (cost=2674.66..2674.68 rows=7 width=949)
        Sort Key: inserted_at DESC
        ->  Index Scan using tasks_user_id_index on queries q0  (cost=0.42..2674.62 rows=7 width=949)
              Index Cond: (user_id = 126)
              Filter: (context = 'http'::query_context)
```

Adding a composite index reduces this down to more or less an
instantaneous lookup:

```
Limit  (cost=11.03..11.03 rows=1 width=1358)
  ->  Sort  (cost=11.03..11.04 rows=2 width=1358)
        Sort Key: inserted_at DESC
        ->  Index Scan using user_context_idx_quereis on queries q0  (cost=0.43..11.02 rows=2 width=1358)
              Index Cond: ((user_id = 1) AND (context = 'http'::query_context))
```

Notice that the first query plan is from the demo host (no index) whereas the
second one is from the attack system (manually created index).
On the attack host the query cost was in the 10s of thousands for
Daniel, resulting in the query taking close to a minute to execute
(exceeding the 15s timeout we had set) causing the request to fail.