Rename data_sources to reflect what it contains
Lacking proper typing, it's sort of hard to know
what a variable contains. In this case we had a
variable called data sources that didn't actually
contain the data source themselves at all.
It was unneccessarily complex to understand
the flow given the assumptions that come with
the variable name.