Clarify logic and move from 1+N queries to 1
The logic was somewhat convoluted to follow.
Particularly I felt it was unintuitive that selected_tables
returned all tables when none were selected.

This commit makes the logic more explicit, and names
the variables to indicate what they actually hold.

This commit also moves from requiring 1+N database
queries to render the page to 1. This makes a significant
change to initial load times which was rather poor for
systems with a larger number of data sources such as
the demo system.