Delay deleting of explorer results
Truth be told I haven't yet fully validated that this is
the cause of the problem of the vanishing results.
However I don't see how it would otherwise happen.

It is not unheard of that the cloaks send some sort of
intermediate result back to the air. If this result does not
contain all the tables in the database (because of a db
glitch causing the connection to fail, causing the set of
tables that's returned to be empty), then all the analyses
results, along with the fact that the tables should be analyzed
at all was forgotten!

This commit adds a soft delete. Soft deleted results
are permanently deleted every 10 days.

Closes #4805