Unify footers
This has been something that annoyed me for a long time,
so it feels great to have this out of the system :)

Our footers were all over the place. There was a total lack
of consistency with how much space they had around them, and
where they were placed.

This commit changes the following:

- it makes the vertical spacing around the footer consistent across
  the query page, the main non-admin pages and the admin pages
- it gives the footers a tiny bit more vertical breathing room
  to make the admin footer slightly less cramped
- it makes the border above the footer slightly lighter and less
  in your face for a slightly more friendly experience

One thing which I couldn't find an obvious way to fix was that
the footer now has some padding on the query page. This results in
the border on the top ending 15px too soon and starting 15px too late
on the query page. That being said it actually looks quite nice,
so I guess it's not a big deal. All in all it's still miles more
consistent than it was in the past.