Footer alignment improvements
This improves the footer alignment further.
The previous change made the height and formatting consisten. However the
center alignment only really made sense on the screens that had the
content center aligned too (such as the data sources overview page, or
the privacy policy page). Other parts of our application such as the admin
interface, or the query editor are left adjusted, and therefore look
significantly better with the footer being left adjusted too.