Gracefully handle failed type checking
In the user channel we assumed the data source service would always
succesfully typecheck a query, and did not handle any of the potential
errors the service could return. This would in turn crash the whole
channel in case of an error. Such an error happened when a user would
visit the data source editor of a data source that was offline.

In this change we gracefully handle the potential error conditions,
returning an error to the frontend.

Closes #4861