Show login events on status page
Telefonica wanted an overview over login attempts.
Both failed and successful.

I am not at all happy with this outcome. It does show login
attempts, but it doesn't really go all the way. Among other things
they also want to see a log of general system events... Maybe this
could all be combined into one?

What annoys me the most about the current design is that it requires
a new set of service (the login stats gen server). The reason for this
requirement is that the audit log only records failed login attempts for
known and existing users, whereas this log should also show when
someone is bruteforcing the system. Extending the audit log to include
such data as well is not straight forward as all audit log entries are, by
design, associated with a user...