Create system user on demand
The system user should not be possible to delete.
However if it is acidentally deleted (say manually by
a DBA, or programmatically through a test), then it
won't be recreated and the system will break.

The recreation won't happen as the user was created
as part of a database migration which won't run again.

This commit changes the behavior of loading the system
user by ensuring it is created should it be missing.