Links go where you expect them
We had this auto-magic routing where the admin link
would take you one place if there were warnings, and
another if there were none. This could be slightly annoying
and confusing. This has now been removed in its entirety
and does now require that an admin who wants to see what
the warnings are needs to click on "admin > warnings" instead.