Note that deleting disabled users might take a while
The deletion happening asynchronously means that most likely, if there
were many disabled users, the deletion can take a while, and will
certainly not be complete by the time the user is redirected back to
the page. This can be confusing. If it's spelled out that things are
slow, then hopefully that helps.