Report per user query events based on audit log
The previous implementation returned the number of queries someone had
run based on the queries table. This was flawed. It would not count
failed queries, and furthermore someone could doctor their stats by
deleting their past queries. This status page widget is meant to give
an administrator a heads-up in case there is someone with malicious
intent running abnormally many queries, and if this stat can be
controlled or influenced by said malicious user, then it's entirely
useless.